% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidybrowse_deps}
\alias{tidybrowse_deps}
\title{List all tidybrowse dependencies}
\usage{
tidybrowse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidybrowse packages.}
}
\description{
List all tidybrowse dependencies
}
