% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidybrowse_packages}
\alias{tidybrowse_packages}
\title{List all packages in the tidybrowse}
\usage{
tidybrowse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidybrowse in the list?}
}
\description{
List all packages in the tidybrowse
}
\examples{
tidybrowse_packages()
}
