% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidybrowse_update}
\alias{tidybrowse_update}
\title{Update tidybrowse packages}
\usage{
tidybrowse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
tidybrowse packages.}
}
\description{
This will check to see if all tidybrowse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
tidybrowse_update()
}
}
