% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidybrowser_deps}
\alias{tidybrowser_deps}
\title{List all tidybrowser dependencies}
\usage{
tidybrowser_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidybrowser packages.}
}
\description{
List all tidybrowser dependencies
}
