% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidybrowser_packages}
\alias{tidybrowser_packages}
\title{List all packages in the tidybrowser}
\usage{
tidybrowser_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidybrowser in the list?}
}
\description{
List all packages in the tidybrowser
}
\examples{
tidybrowser_packages()
}
