% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidybrowser_update}
\alias{tidybrowser_update}
\title{Update tidybrowser packages}
\usage{
tidybrowser_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
tidybrowser packages.}
}
\description{
This will check to see if all tidybrowser packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
tidybrowser_update()
}
}
